clean up docs site functions
