+++ scripts/generate-docs-index/utils/format-filename.js
@@ -7,4 +7,4 @@ module.exports = function formatFilename(filename) {
   filename = `${filename.charAt(0).toUpperCase()}${filename.slice(1)}`;
   filename = filename.split('-').map(word => `${word.charAt(0).toUpperCase()}${word.slice(1)}`).join(' ');
   return filename.replace('.md', '');
-}
+};
